Split run command into platform specific since it now takes different modules for initialization
