+++ .github/CODEOWNERS
@@ -197,6 +197,7 @@
 # BEGIN COMPONENTS
 /comp @DataDog/agent-shared-components
 /comp/aggregator @DataDog/agent-shared-components
+/comp/checks @DataDog/agent-shared-components
 /comp/core @DataDog/agent-shared-components
 /comp/dogstatsd @DataDog/agent-metrics-logs
 /comp/forwarder @DataDog/agent-shared-components
@@ -209,6 +210,7 @@
 /comp/remote-config @DataDog/remote-config
 /comp/systray @DataDog/windows-agent
 /comp/trace @DataDog/agent-apm
+/comp/checks/agentcrashdetect @DataDog/windows-kernel-integrations
 /comp/core/sysprobeconfig @DataDog/ebpf-platform
 # END COMPONENTS
