+++ cmd/agent/subcommands/run/command.go
@@ -35,7 +35,6 @@ import (
 	"github.com/DataDog/datadog-agent/cmd/manager"
 
 	// checks implemented as components
-	"github.com/DataDog/datadog-agent/comp/checks/agentcrashdetect"
 
 	// core components
 	"github.com/DataDog/datadog-agent/comp/core"
@@ -57,11 +56,9 @@ import (
 	"github.com/DataDog/datadog-agent/comp/metadata/runner"
 	"github.com/DataDog/datadog-agent/comp/ndmtmp"
 	"github.com/DataDog/datadog-agent/comp/netflow"
-	netflowServer "github.com/DataDog/datadog-agent/comp/netflow/server"
 	"github.com/DataDog/datadog-agent/comp/otelcol"
 	otelcollector "github.com/DataDog/datadog-agent/comp/otelcol/collector"
 	"github.com/DataDog/datadog-agent/comp/remote-config/rcclient"
-	comptraceconfig "github.com/DataDog/datadog-agent/comp/trace/config"
 	"github.com/DataDog/datadog-agent/pkg/aggregator"
 	"github.com/DataDog/datadog-agent/pkg/api/healthprobe"
 	"github.com/DataDog/datadog-agent/pkg/autodiscovery/providers"
@@ -151,8 +148,7 @@ func Commands(globalParams *command.GlobalParams) []*cobra.Command {
 				LogParams:            log.LogForDaemon(command.LoggerName, "log_file", path.DefaultLogFile),
 			}),
 			getSharedFxOption(),
-			agentcrashdetect.Module,
-			comptraceconfig.Module,
+			getPlatformModules(),
 		)
 	}
 
@@ -177,7 +173,9 @@ func Commands(globalParams *command.GlobalParams) []*cobra.Command {
 // run starts the main loop.
 //
 // This is exported because it also used from the deprecated `agent start` command.
-func run(log log.Component,
+//
+// run now has different parameters on Linux & Windows
+func commonRun(log log.Component,
 	config config.Component,
 	flare flare.Component,
 	telemetry telemetry.Component,
@@ -193,9 +191,6 @@ func run(log log.Component,
 	cliParams *cliParams,
 	logsAgent util.Optional[logsAgent.Component],
 	otelcollector otelcollector.Component,
-	_ netflowServer.Component,
-	_ agentcrashdetect.Component,
-	_ comptraceconfig.Component,
 ) error {
 	defer func() {
 		stopAgent(cliParams, server)
@@ -242,88 +237,6 @@ func run(log log.Component,
 	}
 
 	return <-stopCh
-}
-
-// StartAgentWithDefaults is a temporary way for other packages to use startAgent.
-// Starts the agent in the background and then returns.
-//
-// @ctxChan
-//   - After starting the agent the background goroutine waits for a context from
-//     this channel, then stops the agent when the context is cancelled.
-//
-// Returns an error channel that can be used to wait for the agent to stop and get the result.
-func StartAgentWithDefaults(ctxChan <-chan context.Context) (<-chan error, error) {
-	errChan := make(chan error)
-
-	// run startAgent in an app, so that the log and config components get initialized
-	go func() {
-		err := fxutil.OneShot(func(
-			log log.Component,
-			config config.Component,
-			flare flare.Component,
-			telemetry telemetry.Component,
-			sysprobeconfig sysprobeconfig.Component,
-			server dogstatsdServer.Component,
-			serverDebug dogstatsdDebug.Component,
-			capture replay.Component,
-			rcclient rcclient.Component,
-			forwarder defaultforwarder.Component,
-			logsAgent util.Optional[logsAgent.Component],
-			metadataRunner runner.Component,
-			sharedSerializer serializer.MetricSerializer,
-			otelcollector otelcollector.Component,
-			_ netflowServer.Component,
-			_ agentcrashdetect.Component,
-			_ comptraceconfig.Component,
-		) error {
-
-			defer StopAgentWithDefaults(server)
-
-			err := startAgent(&cliParams{GlobalParams: &command.GlobalParams{}}, log, flare, telemetry, sysprobeconfig, server, capture, serverDebug, rcclient, logsAgent, forwarder, sharedSerializer, otelcollector)
-			if err != nil {
-				return err
-			}
-
-			// notify outer that startAgent finished
-			errChan <- err
-			// wait for context
-			ctx := <-ctxChan
-
-			// Wait for stop signal
-			select {
-			case <-signals.Stopper:
-				log.Info("Received stop command, shutting down...")
-			case <-signals.ErrorStopper:
-				_ = log.Critical("The Agent has encountered an error, shutting down...")
-			case <-ctx.Done():
-				log.Info("Received stop from service manager, shutting down...")
-			}
-
-			return nil
-		},
-			// no config file path specification in this situation
-			fx.Supply(core.BundleParams{
-				ConfigParams:         config.NewAgentParamsWithSecrets(""),
-				SysprobeConfigParams: sysprobeconfig.NewParams(),
-				LogParams:            log.LogForDaemon(command.LoggerName, "log_file", path.DefaultLogFile),
-			}),
-			getSharedFxOption(),
-			agentcrashdetect.Module,
-			comptraceconfig.Module,
-		)
-		// notify caller that fx.OneShot is done
-		errChan <- err
-	}()
-
-	// Wait for startAgent to complete, or for an error
-	err := <-errChan
-	if err != nil {
-		// startAgent or fx.OneShot failed, caller does not need errChan
-		return nil, err
-	}
-
-	// startAgent succeeded. provide errChan to caller so they can wait for fxutil.OneShot to stop
-	return errChan, nil
 }
 
 func getSharedFxOption() fx.Option {
