+++ comp/checks/agentcrashdetect/agentcrashdetect.go
@@ -2,6 +2,7 @@
 // under the Apache License Version 2.0.
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2021-present Datadog, Inc.
+
 //go:build windows
 
 // Package agentcrashdetect detects agent crashes and reports them
